requirements.txt
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-WTF==1.0.0
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
platformdirs==2.4.1
protobuf==3.19.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
spyne==2.13.16
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
zeep==4.1.0
